build on Windows

+++ tools/protobuf.h
@@ -108,7 +108,7 @@ namespace JASS
 			*/
 			inline static int64_t un_zigzag(uint64_t bits)
 				{
-				return (int64_t)((bits >> 1) ^ -(bits & 0x1));
+				return (int64_t)(bits >> 1) ^ -((int64_t)(bits & 0x1));
 				}
 
 		public:
